Reorganize files and URLs
It occurred to me that I might accidentally use the URL of the original
files unless there was something in the URL which indicated that they
are original. For that reason, originals are now in a directory named
"original" and every other image for blog posts must be accessed from
within a directory named "variants".